+++ CHANGELOG.md
@@ -5,9 +5,16 @@ Version changes are pinned to SDK releases.
 
 ## Unreleased
 
+- Add post only front order type.
+
+## [1.18.0]
+
+- Add sei. ([#339](https://github.com/zetamarkets/sdk/pull/339))
+- Post only front order type. ([#312](https://github.com/zetamarkets/sdk/pull/312))
+
 ## [1.17.2]
 
-- BONK constants ([#337](https://github.com/zetamarkets/sdk/pull/337))
+- BONK constants. ([#337](https://github.com/zetamarkets/sdk/pull/337))
 
 ## [1.17.1]